Move fmanager to separate directory. Use callbacks. Small fix to menu to allow clear finished flag from callback. Selection mode. Status bar. Multiple files move/copy/delete
